refactor: move classe para /contracts
